requirements.txt
asgiref==3.2.10
certifi==2019.9.11
confusable-homoglyphs==3.2.0
Django==3.0.8
django-address==0.2.4
django-appconf==1.0.4
django-avatar==5.0.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-grappelli==2.14.2
django-photologue==3.12
django-profile==0.0.3
django-registration==3.1
django-sortedm2m==3.0.2
ExifRead==2.2.1
Pillow==7.2.0
pipenv==2018.11.26
pytz==2020.1
South==1.0.2
sqlparse==0.3.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
asgiref==3.2.10
bootstrap4==0.1.0
certifi==2019.9.11
confusable-homoglyphs==3.2.0
Django==3.0.8
django-address==0.2.4
django-appconf==1.0.4
django-avatar==5.0.0
django-crispy-forms==1.9.2
django-crum==0.7.7
django-debug-toolbar==2.2
django-grappelli==2.14.2
django-photologue==3.12
django-profile==0.0.3
django-registration==3.1
django-sortedm2m==3.0.2
ExifRead==2.2.1
Pillow==7.2.0
pipenv==2018.11.26
pytz==2020.1
South==1.0.2
sqlparse==0.3.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
